requirements.txt
opentele==1.15.1
Telethon==1.28.5
python-dotenv~=1.0.0